Django/fullstack_developer_capstone/server/djangoapp/microservices/app.py
from flask import Flask
from nltk.sentiment import SentimentIntensityAnalyzer
import json
import os
import nltk

app = Flask("Sentiment Analyzer")

# Download required NLTK data
try:
    nltk.data.find('vader_lexicon')
except Exception:
    nltk.download('vader_lexicon')

sia = SentimentIntensityAnalyzer()


@app.get('/')
def home():
    return "Welcome to the Sentiment Analyzer. \
    Use /analyze/text to get the sentiment"


@app.get('/analyze/<input_txt>')
def analyze_sentiment(input_txt):

    scores = sia.polarity_scores(input_txt)
    print(scores)
    pos = float(scores['pos'])
    neg = float(scores['neg'])
    neu = float(scores['neu'])
    res = "positive"
    print("pos neg nue ", pos, neg, neu)
    if (neg > pos and neg > neu):
        res = "negative"
    elif (neu > neg and neu > pos):
        res = "neutral"
    res = json.dumps({"sentiment": res})
    print(res)
    return res


if __name__ == "__main__":
    port = int(os.environ.get('PORT', 5000))
    app.run(host='0.0.0.0', port=port, debug=False)
